Moved DOMAIN to const.py
