+++ homeassistant/components/wilight/config_flow.py
@@ -3,11 +3,11 @@ from urllib.parse import urlparse
 
 import pywilight
 
-from . import DOMAIN
-
 from homeassistant.components import ssdp
 from homeassistant.config_entries import CONN_CLASS_LOCAL_PUSH, ConfigFlow
 from homeassistant.const import CONF_HOST
+
+from .const import DOMAIN
 
 CONF_SERIAL_NUMBER = "serial_number"
 CONF_MODEL_NAME = "model_name"
