+++ homeassistant/components/wilight/const.py
@@ -1,0 +1,3 @@
+"""Const for WiLight integration."""
+
+DOMAIN = "wilight"
