+++ homeassistant/components/wilight/cover.py
@@ -15,7 +15,8 @@ from homeassistant.components.cover import ATTR_POSITION, CoverEntity
 from homeassistant.config_entries import ConfigEntry
 from homeassistant.core import HomeAssistant
 
-from . import DOMAIN, WiLightDevice
+from . import WiLightDevice
+from .const import DOMAIN
 
 
 async def async_setup_entry(
